Added table for data fusion

+++ src/main/java/de/uni_mannheim/informatik/wdi/datafusion/usecase/companies/fusers/ProfitFuser.java
@@ -5,31 +5,33 @@ import de.uni_mannheim.informatik.wdi.datafusion.FusedValue;
 import de.uni_mannheim.informatik.wdi.datafusion.RecordGroup;
 import de.uni_mannheim.informatik.wdi.datafusion.conflictresolution.meta.FavourSources;
 import de.uni_mannheim.informatik.wdi.datafusion.usecase.companies.FusableCompany;
+import de.uni_mannheim.informatik.wdi.datafusion.usecase.companies.conflictresolution.Max;
 
-public class ProfitFuser extends AttributeValueFuser<Long, FusableCompany>  {
+public class ProfitFuser extends AttributeValueFuser<Double, FusableCompany>  {
 
 	public ProfitFuser() {
+		//super(new Max<FusableCompany>());
 		super(new FavourSources<>());
 	}
 	
 	@Override
-	protected Long getValue(FusableCompany record) {
-		return record.getProfit();
+	protected Double getValue(FusableCompany record) {
+		return (double) record.getProfit();
 	}
 	
 	@Override
 	public void fuse(RecordGroup<FusableCompany> group, FusableCompany fusedRecord) {
 		// get the fused value
-		FusedValue<Long, FusableCompany> fused = getFusedValue(group);
+		FusedValue<Double, FusableCompany> fused = getFusedValue(group);
 		
 		//System.out.println("RevenueFuser: " + fused.getValue() + fused.getOriginalIds());
 		
-		Long result = fused.getValue();
+		Double result = fused.getValue();
 		if (result == null)
-			result = 0L;
+			result = 0.0;
 		
 		// set the value for the fused record
-		fusedRecord.setProfit(result);
+		fusedRecord.setProfit(result.longValue());
 		
 		// add provenance info
 		fusedRecord.setAttributeProvenance(FusableCompany.PROFIT, fused.getOriginalIds());
